requirements.txt
absl-py==2.0.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.85
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
etils==1.5.2
filelock==3.13.1
flax==0.7.5
fonttools==4.46.0
fsspec==2023.12.1
gast==0.5.4
gatspy==0.3
gif==23.3.0
idna==3.6
importlib-resources==6.1.1
jax==0.4.21
jaxlib==0.4.21
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
neurokit2==0.2.7
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.7
packaging==23.2
pandas==2.1.3
pathlib2==2.3.7.post1
Pillow==10.1.0
protobuf==4.25.1
py-ecg-detectors==1.3.5
pyarrow==15.0.0
pycparser==2.21
Pygments==2.17.2
pymssql==2.2.11
pyparsing==3.1.1
PyPrind==2.11.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
scikit-learn==1.4.1.post1
scipy<1.10
seaborn==0.13.0
six==1.16.0
soundfile==0.12.1
sympy==1.12
tensorflow-probability==0.23.0
tensorstore==0.1.51
threadpoolctl==3.3.0
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
triton==2.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wfdb==4.1.2
xgboost==2.0.3
zipp==3.17.0
pytorch_forecasting
lxml
statsmodels
